26 of 100 Days of Swift
Completed the last challenge for Project 3 which was to add a
right bar button in "Guess the Flag" to cause a pop-up to show
the current score. I also started project 4, "Easy Browser," which
is a simple introduction to WebKit's WKWebView. I also modified
the tutorial's code to use a dictionary of webstites and iterate
over them to add to the "Open" controller options instead of
manually listing all of the websites.