26 of 100 Days of Swift (cont.)
I continued working on the browser project in the evening. I add
a tool bar to the bottom for backward/forward navigation, a
loading progress bar, a share button (my own addition) and a
refresh button. There is also a feature that blocks unknown
hosting names and an alert pops up if a clicked link is blocked.