28 of 100 Days of Swift
I learned how to create a layout using Auto Layout and
constraints. The example uses the safe area constrains on the
leading edge (where the text is) but not the trailing edge (where
there is not text to protect). I learned how to request and parse
a JSON file in the "Whitehouse Petitions" project. I completed the
challenges: 1) add a credits pop-up, 2) add more information to
each petition detailed view, 3) allow for filtering by the
petition titles.