29 of 100 Days of Swift
I finished the challeneges, which included a pop-up when a wrong
answer is submitted. I included two buttons, one to clear the
current answer, and another to continue with the current letters
entered. There are also point removed for wrong answers, which
required me to change the criteria for checking if the level was
complete.