37 of 100 Days of Swift
I got a lot of practice making a layout in code - the tutorial
created it in the IB, but I found it easier, and more adaptable
to build it in `viewDidLoad()`. I did run into a problem with
showing the label in a `UIButton`, but figured it out by setting
the `UIButton(.roundedRect)` I got a refresher on loading images
from the library. Finally, I learned about the basics of Core
Images.