39 of 100 Days of Swift
I completed the "Whack-a-Penguin" app, except for the challenges
at the end of the section. I learned a few more methods available
for `SKActions` and how to think about the gameplay. I have also
read a few posts on a blog about best practices in iOS development,
includeing  [Passing Data Between View Controllers in iOS: The
Definitive Guide](https://matteomanferdini.com/how-ios-view-
controllers-communicate-with-each-other/). It was really
interesting and useful.