41 of 100 Days of Swift
It was a  pretty simple project that taught me about MapKit. I
learned about how to add pins and modify their pop-ups when tapped,
including adding text and buttons (along with their selector
functions). I also modified the pin colors to be randomly chosen.
There is a navigation button to change the type of map view.