43 of 100 Days of Swift
I continued onto the next project, "Project 19 JavaScript
Injection." The main focus of the app is the extension, but I went
ahead and made the app portion a little more interesting by hosting
a disco! So far, we set up the framework for the Safari extension,
loading the first item, but the extension doesn't do anything.