44 of 100 Days of Swift
I finished the app such that JavaScript code is run in a Safari
extension. I also completed the challenge problems which included
adding a button to add example code. Also, the data for each
website is saved and automatically loaded the next time the
extension is used on the same webpage.