50 of 100 Days of Swift
I completed 2 of the 3 challenges in the end of the Swifty Ninja
project. The first was to remove absolute frame values with those
that adjust to the frame of the device. The second was to show a
Game Over label. I also completed the playground to learn about
strings and `NSAttributedString`s.