51 of 100 Days of Swift
I created an app that loads images from the Photos library and
shares them to all connected devices. In the example, I create a
network between the iOS Simulator and my iPhone 6S, using the
iPhone as the host. I then load an image from the library, and
it instantly appears on the simulator. The same can be done in
other direction, too.