53 of 100 Days of Swift
I continued working on the Plant tracker app for a few hours today.
I was able to add in the three information views to the detail view
controller for the Library tab. After fixing a few constraints and
the size of the main scroll view's content, the scrolling of the
main scroll view and the subview tables was intuitive and
satistfactory.