53 of 100 Days of Swift (cont)
I finished two of the three challenges for the Selfie Share app
(the third was much more invovled than I really desired). The first
challenge was to display an alert when someone disconnects: this
was implemented in the switch-case statement in
`session(:peer:didChange)`. The second challenge was to add a
button to display all phones connected to the session: if there
were multiple connected peers, an additional button was available
in the `+` alert controller that would present another alert with
effectively the following expression
`[peerID.displayName].joined(", ")`.