54 of 100 Days of Swift
I started the Marble Maze game app. I have completed most of the
set-up including loading level one, adding physics bodies to all
neccesary components, and a player that can roll around in response
to touch on the simulator or tilt on an iPad. I got a quick
introduction to `CoreMotion`, too.