61 of 100 Days of Swift
I did a whole bunch on the Plant Tracker iOS app. The main success
of the day was refactoring how I saved images as they were being
selected. It was a lot more work, but the app is much smoother,
faster, and more stable now, so it was definitely worth it. I also
added an icon image to the Library tab and a new UI feature to the
Header in the Library detail view controller.