67 of 100 Days of Swift
I implemented the interactive zoom transition from the
PhotoZoomAnimator test app this morning - it worked flawlessly from
the start! I finally returned to the Marble Maze game. I finished
up the standard portion of the tutorial to react to the collision
of two nodes. The first task was to figure out which was the marble
and then react to what the other one was (eg. if the other node is
a star, the score increments by one).