70 of 100 Days of Swift
I started Project 28 Secret Swift. The base project was to create a
secret text view that stored the text in Keychain. Only through
biometric identifiaction (either Face ID or Touch ID) could the
text be read. Another important feature was to ensure that the text
was hidden if the app was navigated away from.