71 of 100 Days of Swift
The base app in Project 30 was a simple table view app with
circular icons in each cell. When tapped, a detail view controller
loaded with the full (massive) images and recorded the number of
times it was tapped on. It was a very slow app - for many reasons -
so we used Time Profiler and Acllocations in the Instruments in
Xcode.