82 of 100 Days of Swift
I continued working on the Plant Tracker app. While I was at it, I
\continued refactoring the detail view controller by removing all
table view delegate and data source responsibilities. I also
started work on the dynamic addition and removal of cell to the
general information table view that will be used to all for editing
of the information. I was able to mimic the smooth and simultaneous
row removal and insertion that Overcast uses by calling
`tableView.performBatchUpdates(_ updates: (() -> Void)?,
completion: ((Bool) -> Void)? = nil)`. This may be worth writing a
post for the website about.