89 of 100 Days of Swift
I finished the system for editing the attributes of a plant that
have pre-defined levels. There was one last weird UI bug that I was
able to solve. Now, I just need to write up the documentation for
Leo so he can take a look at the system I implemented.