93 of 100 Days of Swift
I began work on a simple app to help me track the germination of
my seeds. This will just be a simple app for my personal use. I
actually planned out how the framework of the app before beginning
and made a suprising amount of progress for a few minutes of
actual coding. This was a great demonstration of how much I have
learned thus far.