95 of 100 Days of Swift
Still working on the Germination Tracker app. I added a bunch of
functionalities today to allow the user to change all of the
information of the seeds. To change the date, I used a modally-
presented view controller with a date picker and "Enter" and
"Cancel" buttons. The number of seeds sown property is adjusted
with an alert with a text field; the text is then parsed for
integers and stored to the plant. I am still running into a few
problems with the scroll view, but that's expected.