96 of 100 Days of Swift
I made great progress on embedding view controllers to in other
view controllers. I made the notes section for the Germination
Tracker a table view with custom cells. I am currently working on
the interface for adding a new note. I have the rough draft for the
view and ran into a snag on getting it to present. It's great
practice, though.