99 of 100 Days of Swift
I spent a few hours at RightPoint in Boston for the monthly
SwiftCoders meetup. I had a great time and learned a lot from
Kevin, an iOS developer at Chewey. I was able to get his help on
a lot of the bugs that were in my Germination Tracker app.